Initial pass at just implementing the Fiction word game.
Intended to be loaded on a mobile phone and enter the secret word from the card as the lie-brian, then, you can manage the clues on your phone. With a timer on the page.

Clicking or tapping the lies will change which was a lie.

There is no way to reset other than refreshing. You also are expected to hit pause timer.